requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
apscheduler>=3.10.4
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
rich>=13.0.0
garth>=0.4.4
